create useIssue effect
